requirements.txt
openenv-core>=0.2.2
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
httpx>=0.27.0
openai>=1.0.0
